requirements.txt
flask
python-dotenv
gunicorn
requests
mysql-connector-python
dnspython